refactor: refocus window behind spotlight
This commit effectively resolves the bugs detected in the
previous commit. It employs the Objective-C "Copy Window List"
API to retrieve the list of windows displayed on the screen,
locates the spotlight window, and subsequently activates the
next window in the list, excluding any windows designated
as a menubar or always-on-top.